# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: Storage
    description: Object storage upload and serving endpoints.
  - name: users
    description: User management operations
  - name: documents
    description: Document management operations
  - name: chat
    description: AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: File not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current app user profile with role
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload and create a new document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /documents/search:
    get:
      operationId: searchDocuments
      tags: [documents]
      summary: Full-text search across documents
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResultsResponse"
  /documents/{id}/chat:
    post:
      operationId: chatWithDocument
      tags: [chat]
      summary: Chat with AI about a specific document (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageBody"
      responses:
        "200":
          description: SSE stream of AI response chunks
          content:
            text/event-stream: {}
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    AppUser:
      type: object
      required: [id, replitUserId, name, email, role, active, createdAt]
      properties:
        id:
          type: integer
        replitUserId:
          type: string
        name:
          type: string
        email:
          type: [string, 'null']
        role:
          type: string
          enum: [admin, advogado]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateUserBody:
      type: object
      required: [replitUserId, name, role]
      properties:
        replitUserId:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        email:
          type: string
        role:
          type: string
          enum: [admin, advogado]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, advogado]
        active:
          type: boolean
    Document:
      type: object
      required: [id, uploadedBy, title, fileName, storagePath, mimeType, createdAt]
      properties:
        id:
          type: integer
        uploadedBy:
          type: integer
        title:
          type: string
        fileName:
          type: string
        storagePath:
          type: string
        mimeType:
          type: string
        hasExtractedText:
          type: boolean
        createdAt:
          type: string
          format: date-time
        uploaderName:
          type: string
    CreateDocumentBody:
      type: object
      required: [title, fileName, storagePath, mimeType, extractedText]
      properties:
        title:
          type: string
          minLength: 1
        fileName:
          type: string
          minLength: 1
        storagePath:
          type: string
          minLength: 1
        mimeType:
          type: string
          minLength: 1
        extractedText:
          type: string
    DocumentListResponse:
      type: object
      required: [documents, total, page, totalPages]
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    SearchResult:
      type: object
      required: [id, title, fileName, snippet, rank, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        fileName:
          type: string
        snippet:
          type: string
        rank:
          type: number
        createdAt:
          type: string
          format: date-time
        uploaderName:
          type: string
    SearchResultsResponse:
      type: object
      required: [results, total, page, totalPages, query]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
        query:
          type: string
    ChatMessageBody:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatHistoryMessage"
    ChatHistoryMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string